Add in ability to load CanOpenNode xml files, this is far more useful than pure eds as it contains much richer info.
Still need to be able to read/write eds though, and currently its two separate classes, not sure of best way to resolve.
Moving it all to use the xml generated classes has a certain logic but we loose some of the finesse of type checking the enums
and everything becomes strings we are passing around.