Subobject override hack
Add in the ability to manually override the subobject count. The reason for this can be demonstrated in CanOpenNode eeprom.c and the access function for OD 0x1011 this function directly accepts the OD access but also takes subindex from 0x01 to 0x7f to activate different features. It does not require an array this size it only needs the Object Dictionary to allow writes to subindex 0x7f for this to work as the access function is invoked and the default OD storage functions are not.

As i'm using 1011 sub 0x7f, manually adding this every time i regenerate was painful. And rather than make it a specific 1011 hack i've made it a general override so it could be used for other access functions in the same way, if required.