First pass work at #48
I think the following now works in both eds and xml

Product number
Vendor number
Description
Creation Date/Time
Modification Date/Time
Modified by

New fields in xml have appropriate uint/byte types not string. String was used for compatibility with CanOpen Nodes original OD editor

the following caveats exist with date time :

In the EDS file the storage format is on two lines one for date and one for time, the format MUST be  DS306 format that is
time- h:mmtt
date- MM-dd-yyyy

XML also stores in same format one field for time one for date

Internally it is stored as a datetime.

When you type and update it uses the DateTime.parse() function so it should support normal local datetime input.

It may crash, it attempts to always parse in a try() but we will see