Add a trap in the PDO update to ensure that if an object that is PDo mapped gets changed this is caught and not used to generate bad PDO data for export.
See bug #56

I am not declaring this a total fix as it should be able to sort itself out. However to do this is requiring more and more "fudging" this is because the PDO generation code needs removing from the GUI and placing in libedssharp where it can look after itself. So i'm adding this trap in the GUI to warn the user there is an issue. To fix the issue from a user perspective, just click on the drop down PDO mapping combo on the TX/RX pdo tab and it will force a rebuild of the PDO mapping.