#104 Add TIME_OF_DAY support
Add typedefs in a #ifndef to avoid an issue if/when these get added to the CO_DRIVER.h file
Add specific paths for formatting TIME_OF_DAY when writing CO_OD.c/h

Currently it only initializes the default/first member of union. So only a single value is accepted. This is not perfect as it
would be more human readable if we could specify milliseconds and days as separate entries but that really requires a GUI change as well.

I've not fully tested this yet but it should be good.